Refactored for simiplicity; eventhorizon -> connect
