+++ Source/Tooling/CLI/EventHorizon/EventHorizonConsents.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Dolittle. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Collections.Generic;
+
+namespace Dolittle.Vanir.CLI.EventHorizon
+{
+    public class EventHorizonConsents : Dictionary<Guid, EventHorizonConsent>
+    {
+    }
+}
